Introduce a separate wrapper type
