+++ src/core/types.rs
@@ -3,9 +3,11 @@ mod paginated_list;
 mod pagination_options;
 mod raw_attributes;
 mod timestamps;
+mod url_encoded_vec;
 
 pub use api_key::*;
 pub use paginated_list::*;
 pub use pagination_options::*;
 pub use raw_attributes::*;
 pub use timestamps::*;
+pub(crate) use url_encoded_vec::*;
