+++ src/core/types/url_encoded_vec.rs
@@ -1,0 +1,113 @@
+use std::fmt::Display;
+use std::fmt::Write;
+
+use serde::Serialize;
+use serde::{ser, Serializer};
+
+/// A [`Vec`] that can be URL-encoded.
+#[derive(Debug)]
+pub(crate) struct UrlEncodedVec<T: Display>(Vec<T>);
+
+impl<T> Serialize for UrlEncodedVec<T>
+where
+    T: Display,
+{
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let mut serialized = String::new();
+
+        for (index, item) in self.0.iter().enumerate() {
+            write!(&mut serialized, "{}", item).map_err(|err| {
+                ser::Error::custom(format!("failed to write '{}': {}", item, err))
+            })?;
+
+            if index < self.0.len() - 1 {
+                write!(&mut serialized, ",").map_err(|err| {
+                    ser::Error::custom(format!("failed to write separator: {}", err))
+                })?
+            }
+        }
+
+        serializer.serialize_str(&serialized)
+    }
+}
+
+impl<T> From<Vec<T>> for UrlEncodedVec<T>
+where
+    T: Display,
+{
+    fn from(vec: Vec<T>) -> Self {
+        Self(vec)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use mockito::{self, mock, Matcher};
+    use reqwest::StatusCode;
+    use serde::Serialize;
+
+    use super::*;
+
+    #[tokio::test]
+    async fn it_serializes_a_vec_in_the_query_string() {
+        #[derive(Debug, Serialize)]
+        struct List<'a> {
+            #[serde(rename = "items[]")]
+            pub items: UrlEncodedVec<&'a str>,
+        }
+
+        let _mock = mock("GET", "/")
+            .match_query(Matcher::UrlEncoded(
+                "items[]".to_string(),
+                "one,two,three".to_string(),
+            ))
+            .with_status(200)
+            .create();
+
+        let client = reqwest::Client::new();
+
+        let response = client
+            .get(&mockito::server_url())
+            .query(&List {
+                items: UrlEncodedVec(vec!["one", "two", "three"]),
+            })
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK)
+    }
+
+    #[tokio::test]
+    async fn it_serializes_a_vec_in_an_option_in_the_query_string() {
+        #[derive(Debug, Serialize)]
+        struct List<'a> {
+            #[serde(rename = "items[]")]
+            pub items: Option<UrlEncodedVec<&'a str>>,
+        }
+
+        let _mock = mock("GET", "/")
+            .match_query(Matcher::UrlEncoded(
+                "items[]".to_string(),
+                "one,two,three".to_string(),
+            ))
+            .with_status(200)
+            .create();
+
+        let client = reqwest::Client::new();
+
+        let response = client
+            .get(&mockito::server_url())
+            .query(&List {
+                items: Some(UrlEncodedVec(vec!["one", "two", "three"])),
+            })
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK)
+    }
+}
